backend/services/attempt_service/pyproject.toml
[project]
name = "attempt-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.26",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.46.0",
]
